feat(src): Rebuild for Arch Linux
The script has been rewritten to be able to support Arch Linux systems.
The current script hasn't been tested. Some of the packages from the
original script require the AUR to be installed, meaning they require a
pacman wrapper.

The next step is to implement options for a pacman wrappers. I think the
best option is to just offer pacman, as it is the only wrapper I have
any experience with. After that, there needs to be checkers for options
that require a wrapper to kick a message to the user saying 'no can do'
if they don't have the wrapper. Alongside that, it seems reasonable to
offer the open source variants for applications like VS Code and Google
Chrome, even though I tend not to use them.